some fixes

+++ Projectiles/EnchantedChakram.cs
@@ -11,9 +11,10 @@ namespace Emperia.Projectiles
     {
     	bool returnTo = false;
 		int timerLeft = 0;
+		private Vector2 direction;
         public override void SetDefaults()
         {
-            projectile.name = "Element Ball";
+            projectile.name = "Enchanted Chakram";
             projectile.width = 8;
             projectile.height = 8;
             projectile.friendly = true;
@@ -75,59 +76,21 @@ namespace Emperia.Projectiles
             }
             else
             {
-                float num1 = player.position.X + (float)(player.width / 2);
-                float num2 = player.position.Y + (float)(player.height / 2);
-                float num3 = Math.Abs(projectile.position.X + (float)(projectile.width / 2) - num1) + Math.Abs(projectile.position.Y + (float)(projectile.height / 2) - num2);
-
-                if (!targetPlayer)
-                {
-                    ProjCenterX = num1;
-                    ProjCenterY = num2;
-                    targetPlayer = true;
-                }
-                if (targetPlayer)
-                {
-                    float num4 = Main.rand.Next(30, 43);
-                    Vector2 vector35 = new Vector2(projectile.position.X + (float)projectile.width * 0.5f, projectile.position.Y + (float)projectile.height * 0.5f);
-                    float num5 = ProjCenterX - vector35.X;
-                    float num6 = ProjCenterY - vector35.Y;
-                    float num7 = (float)Math.Sqrt((double)(num5 * num5 + num6 * num6));
-                    num7 = num4 / num7;
-                    num5 *= num7;
-                    num6 *= num7;
-                    projectile.velocity.X = (projectile.velocity.X * 20f + num5) / 21f;
-                    projectile.velocity.Y = (projectile.velocity.Y * 20f + num6) / 21f;
-                    return;
-                }
-
-
+                direction = player.Center - projectile.Center ;
+				direction.Normalize();
+				projectile.velocity = direction * 12f;
             }
             if (timerLeft > 120)
             {
                 returnTo = true;
             }
-            if (projectile.position.X > player.position.X - 20 && projectile.position.X < player.position.X + 20)
-            {
-                despawnCondition1 = true;
-            }
-            else
-            {
-                despawnCondition1 = false;
-            }
-            if (projectile.position.Y > player.position.Y - 50 && projectile.position.X < player.position.Y + 50)
-            {
-                despawnCondition2 = true;
-            }
-            else
-            {
-                despawnCondition2 = false;
-            }
-            if (despawnCondition1 && despawnCondition2)
-            {
-                if (returnTo)
-                    projectile.Kill();
-            }
-            projectile.rotation += 0.25f;
+			
+            if (projectile.position.X > player.position.X - 20 && projectile.position.X < player.position.X + 20 && projectile.position.Y > player.position.Y - 20 && projectile.position.Y < player.position.Y + 20 && returnTo)
+			{
+				projectile.Kill();
+			}
+			
+            projectile.rotation += 0.15f;
         }
         public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
         {
